+++ metadata.rb
@@ -7,4 +7,5 @@ long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version          '0.7.0'
 
 supports 'ubuntu'
-supports 'centos', 'redhat'
+supports 'centos'
+supports 'redhat'