+++ CHANGELOG.md
@@ -6,6 +6,9 @@ v0.7.0 (2015-04-10)
 - Fixed a bunch of rubocop offenses
 - Deferred some rubocop offenses (detailed in .rubocop_todo.yml)
 - added overwrite_original method that copies the temporary file to the original
+- [#25](https://github.com/someara/line-cookbook/issues/25)
+  moved tester recipe to a standalone cookbook
+  chefspec changed to use the test cookbook and ServerRunner
 
 v0.6.1 (2015-02-24)
 --------------------